Reset voltage works.
There were many changes here. I was using the wrong type (long int instead of uint64_t) for voltage, so even setting the voltage in the first place did not work properly. I added '--read', '--reset' and '--set <x>' as command line arguments, for testing only. I also tried to do something about the unknown number of arguments, but that works as it is (with no arguments), and hasn't been tested at all.